+++ github/resource_github_team_test.go
@@ -151,3 +151,61 @@ func TestAccGithubTeamRemovesDefaultMaintainer(t *testing.T) {
 	})
 
 }
+
+func TestAccGithubTeamUpdateName(t *testing.T) {
+	randomID := acctest.RandStringFromCharSet(5, acctest.CharSetAlphaNum)
+
+	t.Run("marks the slug as computed when the name changes", func(t *testing.T) {
+
+		config := fmt.Sprintf(`
+			resource "github_team" "test" {
+				name         = "tf-acc-%s"
+			}
+		`, randomID)
+
+		configUpdated := fmt.Sprintf(`
+			resource "github_team" "test" {
+				name         = "tf-acc-updated-%s"
+			}
+
+			resource "github_team" "other" {
+				name         = "tf-acc-other-%s"
+				description  = github_team.test.slug
+			}
+		`, randomID, randomID)
+
+		testCase := func(t *testing.T, mode string) {
+			resource.Test(t, resource.TestCase{
+				PreCheck:  func() { skipUnlessMode(t, mode) },
+				Providers: testAccProviders,
+				Steps: []resource.TestStep{
+					{
+						Config: config,
+						Check: resource.ComposeTestCheckFunc(
+							resource.TestCheckResourceAttr("github_team.test", "slug", fmt.Sprintf("tf-acc-%s", randomID)),
+						),
+					},
+					{
+						Config: configUpdated,
+						Check: resource.ComposeTestCheckFunc(
+							resource.TestCheckResourceAttr("github_team.other", "description", fmt.Sprintf("tf-acc-updated-%s", randomID)),
+						),
+					},
+				},
+			})
+		}
+
+		t.Run("with an anonymous account", func(t *testing.T) {
+			t.Skip("anonymous account not supported for this operation")
+		})
+
+		t.Run("with an individual account", func(t *testing.T) {
+			t.Skip("individual account not supported for this operation")
+		})
+
+		t.Run("with an organization account", func(t *testing.T) {
+			testCase(t, organization)
+		})
+
+	})
+}